hello_world/requirements.txt
requests
Jinja2==3.1.2
Flask==2.2.2
flask-lambda-python36
aws-wsgi